% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcom_functions.R
\name{cons_comp_analysis}
\alias{cons_comp_analysis}
\title{Run Components Analysis}
\usage{
cons_comp_analysis(data, outcome_state = "b_state", pred_trait = c("P",
  "E"), pred_state = c("s_state"), single = TRUE, id = "id")
}
\arguments{
\item{data}{The data frame to analyze}

\item{outcome_state}{The outcome or dependent variable that is predicted (the average level-1 variable)}

\item{pred_trait}{The predictor variables on level-2}

\item{pred_state}{The predictor variable on level-1}

\item{id}{The id/group variable of the participants}
}
\value{
Returns a list with two elements:
\itemize{
 \item{1}{The results of the components analysis}
 \item{2}{The explained variance per state}
 \item{3}{The results of the dominance analyses}
}
}
\description{
Run Components Analysis
}
\examples{
simulated_data <- sim_data()
data <- simulated_data
data <- cons_comp_analysis(simulated_data)
}

