% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcom_functions.R
\name{plot_com}
\alias{plot_com}
\title{Plot results of components analysis}
\usage{
plot_com(params_df, ...)
}
\arguments{
\item{params_df}{The results of the function, which is a list with to entries.}

\item{...}{Further arguments passed on to plot_polygons}
}
\value{
Returns a plot with the components
}
\description{
Plot results of components analysis
}
\examples{
## note defined yet
}

