% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcom_functions.R
\name{plot_dominance}
\alias{plot_dominance}
\title{Plot Results from Dominance Analyses}
\usage{
plot_dominance(dominance, line_names = c("P", "state", "E"), ...)
}
\arguments{
\item{dominance}{The dominance results from cons_comp_analysis}

\item{line_names}{The names of the lines. Default is P, state, and E}

\item{...}{Further parameters passed on to plot()}
}
\value{
Returns a plot with the results of the dominance analyses
}
\description{
Plot Results from Dominance Analyses
}
\examples{
data <- sim_data()
estimates <- cons_comp_analysis(data, single=FALSE)
plot_dominance(estimates)
}

