% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcom_functions.R
\name{plot_polygons}
\alias{plot_polygons}
\title{Plot Stacked Areas}
\usage{
plot_polygons(data, poly.border.col = "grey10", poly.fill.col = c("white",
  "white", "white", "white", "#F8CBAE", "#FDF2CA", "#9DC3E6"),
  letters = c("P,S,E", "S,E", "P,E", "P,S", "E", "S", "P"),
  letter.col = "black", letters.right = FALSE, ...)
}
\arguments{
\item{data}{The data frame that needs to be plotted. Should only contain the variables that should appear in the plot.}

\item{poly.border.col}{The color of the border of the stacked areas.}

\item{poly.fill.col}{The fill of the stacked areas. The areas are colored from top to bottom (i.e. the highest one first).}

\item{letters}{The letters that should be placed in the areas. If NULL, the names of the data is used.}

\item{letter.col}{The colors of the letters that are plotted.}

\item{...}{Further arguments passed on to the plot function plot(...)}

\item{letters.right.}{Logical, default is FALSE. If TRUE, the letters are set to the right of the plot.}
}
\value{
Plots areas that are scaled to 100%.
}
\description{
Plot Stacked Areas
}
\examples{
simulated_data <- sim_data()
comp_list <- cons_comp_analysis(simulated_data)
plot_data <- comp_list[[1]]
plot_polygons(plot_data, letters.right = TRUE)
}

