% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcom_functions.R
\name{plot_com}
\alias{plot_com}
\title{Plot results of components analysis}
\usage{
plot_com(components, rsquared, ...)
}
\arguments{
\item{...}{Further arguments passed on to plot_polygons}

\item{params_df}{The results of the function, which is a list with to entries.}
}
\value{
Returns a plot with the components
}
\description{
Plot results of components analysis
}
\examples{
data <- sim_data()
comp_list <- cons_comp_analysis(simulated_data)
plot_com(components = comp_list[[1]], rsquared = comp_list[[2]])
}

